refactor: move JSONSchemaContext to jsonSchemaTypes
